Moved Translator child objects to individual files
